+++ src/pss.rs
@@ -105,7 +105,7 @@ impl SignatureScheme for Pss {
         verify(
             pub_key,
             hashed,
-            &BoxedUint::from_be_slice(sig, sig.len() * 8)?,
+            &BoxedUint::from_be_slice(sig, sig.len() as u32 * 8)?,
             sig.len(),
             &mut *self.digest,
             self.salt_len,
