+++ src/traits/keys.rs
@@ -36,14 +36,14 @@ pub trait PublicKeyPartsNew {
 
     fn n_params(&self) -> BoxedResidueParams;
 
-    fn n_bits_precision(&self) -> usize {
+    fn n_bits_precision(&self) -> u32 {
         self.n().bits_precision()
     }
 
     /// Returns the modulus size in bytes. Raw signatures and ciphertexts for
     /// or by this public key will have the same size.
     fn size(&self) -> usize {
-        (self.n().bits() + 7) / 8
+        (self.n().bits() as usize + 7) / 8
     }
 }
